REFACTOR: Moved expand collapse button to button directory
